+++ src/il2cpptoolkit/Il2CppToolkit.Runtime/Il2CppToolkit.Runtime/ITypeInfoProvider.cs
@@ -1,0 +1,9 @@
+using System;
+using Il2CppToolkit.Injection.Client;
+
+namespace Il2CppToolkit.Runtime;
+
+public interface ITypeInfoProvider
+{
+    bool TryGetTypeInfo(Il2CsRuntimeContext runtime, Type managedType, out Il2CppTypeInfo result);
+}